修改了一些细节

+++ src/LItem.h
@@ -1,11 +1,9 @@
-#include <fstream>
-#include <iostream>
 using namespace std;
 
 class LItem {
     public:
         LItem(int index = -1, float value = -1.0f): index_(index), value_(value) {}
-		void set(int index, double value);
+		void set(int index, float value);
 		int getIndex() const;
 		float getValue() const;
     private:
